'fscanf in filterfile read' (OverlayCommonAlgs-00-04-12)
	* fix filter file fscanf
	* OverlayCommonAlgs-00-04-12

2016-04-29 Andy Haas
	* Allow for 64 bit events numbers (need fix
	https://its.cern.ch/jira/browse/ATEAM-286 )
	* Allow for writing of multiple raw skim files (with
	EventOverlayJobTransforms-00-06-22)
	* OverlayCommonAlgs-00-04-11

2016-04-26 Andy Haas
	* Update 0bias skimming to accont for j40 and noalg prescales, for
	2016 data
	* OverlayCommonAlgs-00-04-10


Former-commit-id: 8ce5645e873990b1ffa2d1d6742c43b9cfbf0cbb

+++ Event/EventOverlay/OverlayCommonAlgs/src/BSFilter.h
@@ -30,7 +30,7 @@ class BSFilter : public AthAlgorithm
     int magic;
     filterinfo(): trig(0),nvtx(0), dt(0.0),magic(99999) {}
   };
-  std::map< int, std::map < int, filterinfo > > filtermap;
+  std::map< int, std::map < uint64_t, filterinfo > > filtermap;
 
   FILE *efile;
   std::string m_EventIdFile;
